refactor: separate PermissionController
